Volodymyr_Stohnii/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
dataclasses
Flask==2.0.3
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyTelegramBotAPI==4.11.0
python-telegram-bot==13.12
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
telebot==0.0.5
tornado==6.1
typing_extensions==4.1.1
tzdata==2023.3
tzlocal==4.2
urllib3==1.26.15
Werkzeug==2.0.3
zipp==3.6.0
